work on casts and usual conv

+++ test/execute/0003-locals1.c
@@ -2,7 +2,7 @@
 int
 main()
 {
-	long long x;
+	int x;
 
 	x = 0;
 	return x;
